+++ ft_bzero.c
@@ -6,15 +6,15 @@
 /*   By: hfegrach <hfegrach@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/30 17:31:32 by hfegrach          #+#    #+#             */
-/*   Updated: 2024/10/31 04:28:12 by hfegrach         ###   ########.fr       */
+/*   Updated: 2024/11/12 13:32:41 by hfegrach         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
